requirements.txt
aiogram==3.19.0
aiosqlite==0.19.0
python-dotenv==1.0.0
streamlit==1.32.0
streamlit-telegram-login==0.1.0 